アドバイスした内容を反映
