+++ app/key.py
@@ -1,0 +1,2 @@
+SECRET_KEY = "g8wkf0pvje6m2unhgt3j"
+SALT = "762ifuw9fj5wxjkeu0dk"
